Resolve multiple runtime errors and circular import
